+++ misc/network_servers/v3_nbd/raw.h
@@ -30,8 +30,8 @@ class raw_disk : public v3_disk {
     raw_disk(string & filename);
 
     off_t get_capacity();
-    int read(unsigned char * buf, unsigned long long offset, int length);
-    int write(unsigned char * buf, unsigned long long offset, int length);
+    unsigned int read(unsigned char * buf, off_t offset, int length);
+    unsigned int write(unsigned char * buf, off_t offset, int length);
 
     void attach();
     void detach();
